allow the client endpoint to .connect() multiple times
I think StreamClientEndpoints should be reusable: you should be able to call
.connect() multiple times, yielding a different Protocol instance each time.
But the TorClientEndpoint constructor creates a single iterator to track the
sequence of possible SOCKS ports we should try, so the second time .connect()
is called, it uses the next item on this list, which will fail.

This patch removes the iterator and just loops over the original list each
time .connect() is called, removing the interaction between calls.